Tiny fixes here and there
Fixed bug in Events::new where the custom data was placed at the
most significant bits, but retrieved from the least significant bits.

In both init and process we need a mutable reference to ControlOps to
be able to call add/remove.

Adapted a few tests from previous event manager to check if it works.

Signed-off-by: Andreea Florescu <fandree@amazon.com>