Add details and remove hacks from 'Simplify Subscribers'
This commit makes things pretty so it's not just a PoC change anymore.

Without wrapping `T` in some shared ownership construct like `Arc/Rc`,
I don't see a way of removing the double indirection to get from
`RawFd` to `T`. And the simplest explanation for that is that the mapping
is multiple to one (multiple `RawFd`s can point to a single `T`).

I think the current double indirection is an acceptable trade-off against
the added code complexity of wrapping `T` (which can already be an
`Arc<Mutex>` for example, but we don't know about it) in an extra
shared-ownership wrapper.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>