create separate abstraciton for ready_list
Instead of operating on a Vec<EpollEvent> when handling the list of
ready events, we can work with the newly added abstration `ReadyList`.
This abstraction should also allow other future improvements (for
example creating an iterator over the ready_list that only returns the
items in the range [0..event_count], where event_count is the number
returned by epoll.wait().

Signed-off-by: Andreea Florescu <fandree@amazon.com>